[FIX] rma: default team and responsible
Having a default responsible value as it was defined doesn't behave
right when the user it's a portal one or the petition comes from a
sudo().

Also, by design it was expected to have a default team if it wasn't
defined, but as it was implemented that wasn't possible.

TT32056