[IMP] rma: finish RMA manually
Sometimes there's no choice to return, refund or replace an RMA for
different reasons. For example, when the customer doesn't want to repair
the product because that would be too expensive. We still want finish
the RMA and document the reasons. This improvement allows it.

TT34164

[UPD] Update rma.pot

[UPD] README.rst

rma 14.0.2.0.0

Update translation files

Updated by "Update PO files to match POT (msgmerge)" hook in Weblate.

Translation: rma-14.0/rma-14.0-rma
Translate-URL: https://translation.odoo-community.org/projects/rma-14-0/rma-14-0-rma/