[FIX] rma: Add location_id field for the expected behavior
The location_id field is important and its value is required, we need to add it
to the wizard (even if it is hidden) so that it is saved and the picking is
created correctly.

Before, the location_id value was only saved (although the field was not in the
wizard) if the user had the "Technical / Manage Multiple Stock Locations" permission.