[IMP] rma: custom route for replace action
This commit adds a new field, rma_out_replace_route_id, to the warehouse.
If this field is set, it specifies a custom route to be used for
RMA replacement actions, allowing separation of RMA replacement orders from
regular deliveries. If the field is not set, the default delivery route will be used.